move android parts together
