+++ android/app/.idea/.name
@@ -1,0 +1,1 @@
+Ruffle2
