+++ android/app/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Mar 24 17:39:17 CET 2022
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.2-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
